emmas comments
