+++ src/layers/Layer.ts
@@ -1,5 +1,4 @@
-import { Scales } from "@/Chart";
-import * as d3 from "@/d3";
+import { LayerArgs, ZoomExtents } from "@/types";
 
 /*
   These are the various different types of layer types
@@ -29,54 +28,6 @@ export enum CustomEvents {
   AnimationStart = "animationstart",
   AnimationEnd = "animationend",
 }
-
-export type XY<T> = { x: T, y: T }
-
-export type Point = XY<number>
-
-/*
-  These are bounds of the svg element
-*/
-export type Bounds = {
-  width: number,
-  height: number,
-  margin: {
-    top: number, bottom: number,
-    left: number, right: number
-  }
-}
-
-export type D3Selection<Element extends d3.BaseType> = d3.Selection<Element, Point, null, undefined>
-
-export type AllOptionalLayers = OptionalLayer<any>;
-
-/*
-  LayerArgs are passed into each Layer in the draw
-  function. This happens at the last step when users
-  append the svg to a base element
-*/
-export type LayerArgs = {
-  // TODO chart id instead
-  id: string,
-  getHtmlId: (layer: LayerType) => string,
-  bounds: Bounds,
-  globals: {
-    animationDuration: number
-  },
-  scaleConfig: {
-    linearScales: XY<d3.ScaleLinear<number, number, never>>,
-    scaleExtents: Scales,
-    lineGen: d3.Line<Point>
-  },
-  coreLayers: {
-    [LayerType.Svg]: D3Selection<SVGSVGElement>,
-    [LayerType.ClipPath]: D3Selection<SVGClipPathElement>,
-    [LayerType.BaseLayer]: D3Selection<SVGGElement>,
-  },
-  optionalLayers: AllOptionalLayers[]
-};
-
-export type ZoomExtents = Partial<XY<[number, number]>>
 
 export abstract class OptionalLayer<Properties = null> {
   abstract type: LayerType;
