+++ src/layers/TracesLayer.ts
@@ -1,5 +1,5 @@
-import { Lines } from "@/Chart";
-import { LayerArgs, LayerType, OptionalLayer } from "./Layer";
+import { LayerArgs, Lines } from "@/types";
+import { LayerType, OptionalLayer } from "./Layer";
 
 export class TracesLayer extends OptionalLayer {
   type = LayerType.Trace;
@@ -14,7 +14,6 @@ export class TracesLayer extends OptionalLayer {
       return layerArgs.coreLayers[LayerType.BaseLayer].append("path")
         .attr("id", `${layerArgs.getHtmlId(LayerType.Trace)}-${index}`)
         .attr("pointer-events", "none")
-        .attr("vector-effect", "non-scaling-stroke")
         .attr("fill", "none")
         .attr("stroke", l.style.color || "black")
         .attr("opacity", l.style.opacity || 1)
